[INLONG-8356][DataProxy] Replace source2 to source
