+++ inlong-dataproxy/conf/dataproxy-mulit-pulsar-http-example.conf
@@ -28,13 +28,11 @@ agent1.channels = ch-msg1 ch-msg2 ch-msg5 ch-msg6
 agent1.sinks = pulsar-sink-msg1 pulsar-sink-msg2 pulsar-sink-msg5 pulsar-sink-msg6
 
 agent1.sources.http-source.channels = ch-msg1 ch-msg2 ch-msg5 ch-msg6
-agent1.sources.http-source.type = org.apache.inlong.dataproxy.http.SimpleHttpSource
-agent1.sources.http-source.message-handler-name = org.apache.inlong.dataproxy.http.SimpleMessageHandler
+agent1.sources.http-source.type = org.apache.inlong.dataproxy.source.SimpleHttpSource
+agent1.sources.http-source.message-handler-name = org.apache.inlong.dataproxy.source.httpMsg.HttpMessageHandler
 agent1.sources.http-source.host = 0.0.0.0
 agent1.sources.http-source.port = 46802
 agent1.sources.http-source.max-msg-length = 524288
-#agent1.sources.http-source.topic = persistent://public/default/dataproxy-default-topic
-agent1.sources.http-source.attr = m=9
 agent1.sources.http-source.connections = 30000
 agent1.sources.http-source.max-threads = 64
 agent1.sources.http-source.receiveBufferSize = 1048576
