+++ inlong-dataproxy/conf/dataproxy-mulit-pulsar-udp-example.conf
@@ -33,8 +33,6 @@ agent1.sources.upd-source.msg-factory-name = org.apache.inlong.dataproxy.source.
 agent1.sources.upd-source.host = 0.0.0.0
 agent1.sources.upd-source.port = 46803
 agent1.sources.upd-source.max-msg-length = 524288
-#agent1.sources.upd-source.topic = persistent://public/default/dataproxy-default-topic
-agent1.sources.upd-source.attr = m=9
 agent1.sources.upd-source.connections = 30000
 agent1.sources.upd-source.max-threads = 64
 agent1.sources.upd-source.receiveBufferSize = 1048576
