+++ inlong-dataproxy/conf/dataproxy.conf
@@ -43,8 +43,6 @@ agent1.sources.tcp-source.msg-factory-name = org.apache.inlong.dataproxy.source.
 agent1.sources.tcp-source.host = localhost
 agent1.sources.tcp-source.port = 46801
 agent1.sources.tcp-source.max-msg-length = 524288
-# agent1.sources.tcp-source.topic = persistent://public/default/dataproxy-default-topic
-agent1.sources.tcp-source.attr = m=9
 agent1.sources.tcp-source.connections = 30000
 agent1.sources.tcp-source.max-threads = 64
 agent1.sources.tcp-source.receiveBufferSize = 8388608
@@ -62,14 +60,12 @@ agent1.sources.tcp-source.max-monitor-cnt=500000
 
 # http-source
 agent1.sources.http-source.channels = ch-msg1 ch-msg2 ch-msg5 ch-msg6
-agent1.sources.http-source.type = org.apache.inlong.dataproxy.http.SimpleHttpSource
-agent1.sources.http-source.message-handler-name = org.apache.inlong.dataproxy.http.SimpleMessageHandler
+agent1.sources.http-source.type = org.apache.inlong.dataproxy.source.SimpleHttpSource
+agent1.sources.http-source.message-handler-name = org.apache.inlong.dataproxy.source.httpMsg.HttpMessageHandler
 # replace it with the ip you specified
 agent1.sources.http-source.host = localhost
 agent1.sources.http-source.port = 46802
 agent1.sources.http-source.max-msg-length = 524288
-#agent1.sources.http-source.topic = persistent://public/default/dataproxy-default-topic
-agent1.sources.http-source.attr = m=9
 agent1.sources.http-source.connections = 30000
 agent1.sources.http-source.max-threads = 64
 agent1.sources.http-source.receiveBufferSize = 1048576
