--- inlong-dataproxy/dataproxy-source/src/main/java/org/apache/inlong/dataproxy/http/MappedContext.java
@@ -1,72 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.inlong.dataproxy.http;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
-
-public class MappedContext implements Context {
-
-    private Map<String, Object> mapContext;
-
-    public MappedContext() {
-        init();
-    }
-
-    @Override
-    public void init() {
-        mapContext = new HashMap<>();
-    }
-
-    @Override
-    public void put(String key, Object value) {
-        mapContext.put(key, value);
-    }
-
-    @Override
-    public Object get(String key) {
-        return mapContext.get(key);
-    }
-
-    @Override
-    public void remove(String key) {
-        mapContext.remove(key);
-    }
-
-    @Override
-    public void clear() {
-        mapContext.clear();
-    }
-
-    @Override
-    public void destroy() {
-        mapContext = Collections.emptyMap();
-    }
-
-    @Override
-    public boolean containsKey(String key) {
-        return mapContext.containsKey(key);
-    }
-
-    @Override
-    public Set<String> keySet() {
-        return mapContext.keySet();
-    }
-}
