+++ inlong-dataproxy/dataproxy-source/src/main/java/org/apache/inlong/dataproxy/source/BaseSource.java
@@ -18,145 +18,220 @@
 package org.apache.inlong.dataproxy.source;
 
 import org.apache.inlong.common.metric.MetricRegister;
-import org.apache.inlong.common.monitor.MonitorIndex;
-import org.apache.inlong.common.monitor.MonitorIndexExt;
+import org.apache.inlong.dataproxy.admin.ProxyServiceMBean;
 import org.apache.inlong.dataproxy.channel.FailoverChannelProcessor;
 import org.apache.inlong.dataproxy.config.CommonConfigHolder;
-import org.apache.inlong.dataproxy.consts.ConfigConstants;
+import org.apache.inlong.dataproxy.config.ConfigManager;
+import org.apache.inlong.dataproxy.config.holder.ConfigUpdateCallback;
+import org.apache.inlong.dataproxy.consts.AttrConstants;
+import org.apache.inlong.dataproxy.metrics.DataProxyMetricItem;
 import org.apache.inlong.dataproxy.metrics.DataProxyMetricItemSet;
+import org.apache.inlong.dataproxy.metrics.audit.AuditUtils;
+import org.apache.inlong.dataproxy.metrics.stats.MonitorIndex;
+import org.apache.inlong.dataproxy.metrics.stats.MonitorStats;
+import org.apache.inlong.dataproxy.source.httpMsg.HttpMessageHandler;
+import org.apache.inlong.dataproxy.utils.AddressUtils;
 import org.apache.inlong.dataproxy.utils.ConfStringUtils;
 import org.apache.inlong.dataproxy.utils.FailoverChannelProcessorHolder;
+import org.apache.inlong.sdk.commons.admin.AdminServiceRegister;
 
 import com.google.common.base.Preconditions;
+import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.group.ChannelGroup;
 import io.netty.channel.group.DefaultChannelGroup;
-import io.netty.util.concurrent.DefaultThreadFactory;
 import io.netty.util.concurrent.GlobalEventExecutor;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.flume.ChannelSelector;
 import org.apache.flume.Context;
+import org.apache.flume.Event;
 import org.apache.flume.EventDrivenSource;
 import org.apache.flume.FlumeException;
 import org.apache.flume.conf.Configurable;
-import org.apache.flume.conf.Configurables;
 import org.apache.flume.source.AbstractSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Constructor;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
 
 /**
- * source base clase
- *
+ * source base class
  */
 public abstract class BaseSource
         extends
             AbstractSource
         implements
+            ConfigUpdateCallback,
+            ProxyServiceMBean,
             EventDrivenSource,
             Configurable {
 
     private static final Logger logger = LoggerFactory.getLogger(BaseSource.class);
 
     protected Context context;
-
-    protected int port;
-
-    protected String host = null;
-
+    // whether source reject service
+    protected volatile boolean isRejectService = false;
+    // source service host
+    protected String srcHost;
+    // source serviced port
+    protected int srcPort;
+    protected String strPort;
+    // message factory name
     protected String msgFactoryName;
-
-    protected String serviceDecoderName;
-
+    // message handler name
     protected String messageHandlerName;
-
+    // source default append attribute
+    protected String defAttr = "";
+    // allowed max message length
     protected int maxMsgLength;
-
+    // whether compress message
     protected boolean isCompressed;
-
-    protected String topic;
-
-    protected String attr;
-
+    // whether filter empty message
     protected boolean filterEmptyMsg;
-
-    private int statIntervalSec;
-
-    protected int pkgTimeoutSec;
-
-    protected int maxConnections = Integer.MAX_VALUE;
-
-    private static final String CONNECTIONS = "connections";
-
-    protected boolean customProcessor = false;
-
-    private DataProxyMetricItemSet metricItemSet;
-
-    /*
-     * monitor
-     */
-    private MonitorIndex monitorIndex;
-
-    private MonitorIndexExt monitorIndexExt;
-
-    /*
-     * netty server
-     */
-
+    // whether custom channel processor
+    protected boolean customProcessor;
+    // max netty worker threads
+    protected int maxWorkerThreads;
+    // max netty accept threads
+    protected int maxAcceptThreads;
+    // max read idle time
+    protected long maxReadIdleTimeMs;
+    // max connection count
+    protected int maxConnections;
+    // reuse address
+    protected boolean reuseAddress;
+    // connect backlog
+    protected int conBacklog;
+    // connect linger
+    protected int conLinger = -1;
+    // netty parameters
     protected EventLoopGroup acceptorGroup;
-
     protected EventLoopGroup workerGroup;
-
-    protected DefaultThreadFactory acceptorThreadFactory;
-
-    protected boolean enableBusyWait = false;
-
     protected ChannelGroup allChannels;
-
     protected ChannelFuture channelFuture;
-
-    private static String HOST_DEFAULT_VALUE = "0.0.0.0";
-
-    private static int maxMonitorCnt = ConfigConstants.DEF_MONITOR_STAT_CNT;
-
-    private static int DEFAULT_MAX_CONNECTIONS = 5000;
-
-    private static int STAT_INTERVAL_MUST_THAN = 0;
-
-    private static int PKG_TIMEOUT_DEFAULT_SEC = 3;
-
-    private static int MSG_MIN_LENGTH = 4;
-
-    private static int MAX_MSG_DEFAULT_LENGTH = 1024 * 64;
-
-    private static int INTERVAL_SEC = 60;
-
-    protected static int BUFFER_SIZE_MUST_THAN = 0;
-
-    protected static int DEFAULT_MAX_THREADS = 32;
-
-    protected static int RECEIVE_BUFFER_DEFAULT_SIZE = 64 * 1024;
-
-    protected static int SEND_BUFFER_DEFAULT_SIZE = 64 * 1024;
-
-    protected static int RECEIVE_BUFFER_MAX_SIZE = 16 * 1024 * 1024;
-
-    protected static int SEND_BUFFER_MAX_SIZE = 16 * 1024 * 1024;
-
-    protected int receiveBufferSize;
-
-    protected int sendBufferSize;
-
-    protected int maxThreads = 32;
-
-    protected int acceptorThreads = 1;
+    // receive buffer size
+    protected int maxRcvBufferSize;
+    // send buffer size
+    protected int maxSendBufferSize;
+    // file metric statistic
+    private MonitorIndex monitorIndex = null;
+    private MonitorStats monitorStats = null;
+    // metric set
+    private DataProxyMetricItemSet metricItemSet;
 
     public BaseSource() {
         super();
         allChannels = new DefaultChannelGroup("DefaultChannelGroup", GlobalEventExecutor.INSTANCE);
+    }
+
+    @Override
+    public void configure(Context context) {
+        logger.info("{} start to configure context:{}.", this.getName(), context.toString());
+        this.context = context;
+        this.srcHost = getHostIp(context);
+        this.srcPort = getHostPort(context);
+        this.strPort = String.valueOf(this.srcPort);
+        // get message factory
+        String tmpVal = context.getString(SourceConstants.SRCCXT_MSG_FACTORY_NAME,
+                ServerMessageFactory.class.getName()).trim();
+        Preconditions.checkArgument(StringUtils.isNotBlank(tmpVal),
+                SourceConstants.SRCCXT_MSG_FACTORY_NAME + " config is blank");
+        this.msgFactoryName = tmpVal.trim();
+        // get message handler
+        tmpVal = context.getString(SourceConstants.SRCCXT_MESSAGE_HANDLER_NAME);
+        if (StringUtils.isBlank(tmpVal)) {
+            tmpVal = SourceConstants.SRC_PROTOCOL_TYPE_HTTP.equalsIgnoreCase(getProtocolName())
+                    ? HttpMessageHandler.class.getName()
+                    : ServerMessageHandler.class.getName();
+        }
+        Preconditions.checkArgument(StringUtils.isNotBlank(tmpVal),
+                SourceConstants.SRCCXT_MESSAGE_HANDLER_NAME + " config is blank");
+        this.messageHandlerName = tmpVal;
+        // get default attributes
+        tmpVal = context.getString(SourceConstants.SRCCXT_DEF_ATTR);
+        if (StringUtils.isNotBlank(tmpVal)) {
+            this.defAttr = tmpVal.trim();
+        }
+        // get allowed max message length
+        this.maxMsgLength = ConfStringUtils.getIntValue(context,
+                SourceConstants.SRCCXT_MAX_MSG_LENGTH, SourceConstants.VAL_DEF_MAX_MSG_LENGTH);
+        Preconditions.checkArgument((this.maxMsgLength >= SourceConstants.VAL_MIN_MAX_MSG_LENGTH
+                && this.maxMsgLength <= SourceConstants.VAL_MAX_MAX_MSG_LENGTH),
+                SourceConstants.SRCCXT_MAX_MSG_LENGTH + " must be in ["
+                        + SourceConstants.VAL_MIN_MAX_MSG_LENGTH + ", "
+                        + SourceConstants.VAL_MAX_MAX_MSG_LENGTH + "]");
+        // get whether compress message
+        this.isCompressed = context.getBoolean(SourceConstants.SRCCXT_MSG_COMPRESSED,
+                SourceConstants.VAL_DEF_MSG_COMPRESSED);
+        // get whether filter empty message
+        this.filterEmptyMsg = context.getBoolean(SourceConstants.SRCCXT_FILTER_EMPTY_MSG,
+                SourceConstants.VAL_DEF_FILTER_EMPTY_MSG);
+        // get whether custom channel processor
+        this.customProcessor = context.getBoolean(SourceConstants.SRCCXT_CUSTOM_CHANNEL_PROCESSOR,
+                SourceConstants.VAL_DEF_CUSTOM_CH_PROCESSOR);
+        // get max accept threads
+        this.maxAcceptThreads = ConfStringUtils.getIntValue(context,
+                SourceConstants.SRCCXT_MAX_ACCEPT_THREADS, SourceConstants.VAL_DEF_NET_ACCEPT_THREADS);
+        Preconditions.checkArgument((this.maxAcceptThreads >= SourceConstants.VAL_MIN_ACCEPT_THREADS
+                && this.maxAcceptThreads <= SourceConstants.VAL_MAX_ACCEPT_THREADS),
+                SourceConstants.SRCCXT_MAX_ACCEPT_THREADS + " must be in ["
+                        + SourceConstants.VAL_MIN_ACCEPT_THREADS + ", "
+                        + SourceConstants.VAL_MAX_ACCEPT_THREADS + "]");
+        // get max worker threads
+        this.maxWorkerThreads = ConfStringUtils.getIntValue(context,
+                SourceConstants.SRCCXT_MAX_WORKER_THREADS, SourceConstants.VAL_DEF_WORKER_THREADS);
+        Preconditions.checkArgument((this.maxWorkerThreads >= SourceConstants.VAL_MIN_WORKER_THREADS),
+                SourceConstants.SRCCXT_MAX_WORKER_THREADS + " must be >= "
+                        + SourceConstants.VAL_MIN_WORKER_THREADS);
+        // get max read idle time
+        this.maxReadIdleTimeMs = ConfStringUtils.getLongValue(context,
+                SourceConstants.SRCCXT_MAX_READ_IDLE_TIME_MS, SourceConstants.VAL_DEF_READ_IDLE_TIME_MS);
+        Preconditions.checkArgument((this.maxReadIdleTimeMs >= SourceConstants.VAL_MIN_READ_IDLE_TIME_MS
+                && this.maxReadIdleTimeMs <= SourceConstants.VAL_MAX_READ_IDLE_TIME_MS),
+                SourceConstants.SRCCXT_MAX_READ_IDLE_TIME_MS + " must be in ["
+                        + SourceConstants.VAL_MIN_READ_IDLE_TIME_MS + ", "
+                        + SourceConstants.VAL_MAX_READ_IDLE_TIME_MS + "]");
+        // get max connect count
+        this.maxConnections = ConfStringUtils.getIntValue(context,
+                SourceConstants.SRCCXT_MAX_CONNECTION_CNT, SourceConstants.VAL_DEF_MAX_CONNECTION_CNT);
+        Preconditions.checkArgument(this.maxConnections >= SourceConstants.VAL_MIN_CONNECTION_CNT,
+                SourceConstants.SRCCXT_MAX_CONNECTION_CNT + " must be >= "
+                        + SourceConstants.VAL_MIN_CONNECTION_CNT);
+        // get connect backlog
+        this.conBacklog = ConfStringUtils.getIntValue(context,
+                SourceConstants.SRCCXT_CONN_BACKLOG, SourceConstants.VAL_DEF_CONN_BACKLOG);
+        Preconditions.checkArgument(this.conBacklog >= SourceConstants.VAL_MIN_CONN_BACKLOG,
+                SourceConstants.SRCCXT_CONN_BACKLOG + " must be >= "
+                        + SourceConstants.VAL_MIN_CONN_BACKLOG);
+        // get connect linger
+        Integer tmpValue = context.getInteger(SourceConstants.SRCCXT_CONN_LINGER);
+        if (tmpValue != null && tmpValue >= 0) {
+            this.conLinger = tmpValue;
+        }
+        // get whether reuse address
+        this.reuseAddress = context.getBoolean(SourceConstants.SRCCXT_REUSE_ADDRESS,
+                SourceConstants.VAL_DEF_REUSE_ADDRESS);
+
+        // get whether custom channel processor
+        this.customProcessor = context.getBoolean(SourceConstants.SRCCXT_CUSTOM_CHANNEL_PROCESSOR,
+                SourceConstants.VAL_DEF_CUSTOM_CH_PROCESSOR);
+        // get max receive buffer size
+        this.maxRcvBufferSize = ConfStringUtils.getIntValue(context,
+                SourceConstants.SRCCXT_RECEIVE_BUFFER_SIZE, SourceConstants.VAL_DEF_RECEIVE_BUFFER_SIZE);
+        Preconditions.checkArgument(this.maxRcvBufferSize >= SourceConstants.VAL_MIN_RECEIVE_BUFFER_SIZE,
+                SourceConstants.SRCCXT_RECEIVE_BUFFER_SIZE + " must be >= "
+                        + SourceConstants.VAL_MIN_RECEIVE_BUFFER_SIZE);
+        // get max send buffer size
+        this.maxSendBufferSize = ConfStringUtils.getIntValue(context,
+                SourceConstants.SRCCXT_SEND_BUFFER_SIZE, SourceConstants.VAL_DEF_SEND_BUFFER_SIZE);
+        Preconditions.checkArgument(this.maxSendBufferSize >= SourceConstants.VAL_MIN_SEND_BUFFER_SIZE,
+                SourceConstants.SRCCXT_SEND_BUFFER_SIZE + " must be >= "
+                        + SourceConstants.VAL_MIN_SEND_BUFFER_SIZE);
     }
 
     @Override
@@ -170,184 +245,100 @@ public abstract class BaseSource
         }
         super.start();
         // initial metric item set
-        String clusterId = CommonConfigHolder.getInstance().getClusterName();
-        this.metricItemSet =
-                new DataProxyMetricItemSet(clusterId, this.getName(), String.valueOf(port));
+        this.metricItemSet = new DataProxyMetricItemSet(
+                CommonConfigHolder.getInstance().getClusterName(), getName(), String.valueOf(srcPort));
         MetricRegister.register(metricItemSet);
-        /*
-         * init monitor logic
-         */
-        monitorIndex = new MonitorIndex("Source", INTERVAL_SEC, maxMonitorCnt);
-        monitorIndexExt = new MonitorIndexExt("DataProxy_monitors#"
-                + this.getProtocolName(), INTERVAL_SEC, maxMonitorCnt);
+        // init monitor logic
+        if (CommonConfigHolder.getInstance().isEnableFileMetric()) {
+            this.monitorIndex = new MonitorIndex(CommonConfigHolder.getInstance().getFileMetricSourceOutName(),
+                    CommonConfigHolder.getInstance().getFileMetricStatInvlSec() * 1000L,
+                    CommonConfigHolder.getInstance().getFileMetricStatCacheCnt());
+            this.monitorIndex.start();
+            this.monitorStats = new MonitorStats(
+                    CommonConfigHolder.getInstance().getFileMetricEventOutName()
+                            + AttrConstants.SEP_HASHTAG + this.getProtocolName(),
+                    CommonConfigHolder.getInstance().getFileMetricStatInvlSec() * 1000L,
+                    CommonConfigHolder.getInstance().getFileMetricStatCacheCnt());
+            this.monitorStats.start();
+        }
         startSource();
+        // register
+        AdminServiceRegister.register(ProxyServiceMBean.MBEAN_TYPE, this.getName(), this);
     }
 
     @Override
     public synchronized void stop() {
         logger.info("[STOP {} SOURCE]{} stopping...", this.getProtocolName(), this.getName());
+        // close channels
         if (!allChannels.isEmpty()) {
             try {
                 allChannels.close().awaitUninterruptibly();
             } catch (Exception e) {
-                logger.warn("Simple Source netty server stop ex, {}", e);
+                logger.warn("Close {} netty channels throw exception", this.getName(), e);
             } finally {
                 allChannels.clear();
             }
         }
-
-        super.stop();
-        if (monitorIndex != null) {
-            monitorIndex.shutDown();
-        }
-        if (monitorIndexExt != null) {
-            monitorIndexExt.shutDown();
-        }
-
+        // close channel future
         if (channelFuture != null) {
             try {
                 channelFuture.channel().closeFuture().sync();
             } catch (InterruptedException e) {
-                logger.warn("Simple Source netty server stop ex, {}", e);
+                logger.warn("Close {} channel future throw exception", this.getName(), e);
+            }
+        }
+        // stop super class
+        super.stop();
+        // stop workers
+        if (this.acceptorGroup != null) {
+            this.acceptorGroup.shutdownGracefully();
+        }
+        if (this.workerGroup != null) {
+            this.workerGroup.shutdownGracefully();
+        }
+        // stop file statistic index
+        if (CommonConfigHolder.getInstance().isEnableFileMetric()) {
+            if (monitorIndex != null) {
+                monitorIndex.stop();
+            }
+            if (monitorStats != null) {
+                monitorStats.stop();
             }
         }
         logger.info("[STOP {} SOURCE]{} stopped", this.getProtocolName(), this.getName());
     }
 
     @Override
-    public void configure(Context context) {
-
-        this.context = context;
-
-        port = context.getInteger(ConfigConstants.CONFIG_PORT);
-
-        host = context.getString(ConfigConstants.CONFIG_HOST, HOST_DEFAULT_VALUE);
-
-        Configurables.ensureRequiredNonNull(context, ConfigConstants.CONFIG_PORT);
-
-        Preconditions.checkArgument(ConfStringUtils.isValidIp(host), "ip config not valid");
-        Preconditions.checkArgument(ConfStringUtils.isValidPort(port), "port config not valid");
-
-        msgFactoryName =
-                context.getString(ConfigConstants.MSG_FACTORY_NAME,
-                        "org.apache.inlong.dataproxy.source.ServerMessageFactory");
-        msgFactoryName = msgFactoryName.trim();
-        Preconditions.checkArgument(StringUtils.isNotBlank(msgFactoryName),
-                "msgFactoryName is empty");
-
-        serviceDecoderName =
-                context.getString(ConfigConstants.SERVICE_PROCESSOR_NAME,
-                        "org.apache.inlong.dataproxy.source.DefaultServiceDecoder");
-        serviceDecoderName = serviceDecoderName.trim();
-        Preconditions.checkArgument(StringUtils.isNotBlank(serviceDecoderName),
-                "serviceProcessorName is empty");
-
-        messageHandlerName =
-                context.getString(ConfigConstants.MESSAGE_HANDLER_NAME,
-                        "org.apache.inlong.dataproxy.source.ServerMessageHandler");
-        messageHandlerName = messageHandlerName.trim();
-        Preconditions.checkArgument(StringUtils.isNotBlank(messageHandlerName),
-                "messageHandlerName is empty");
-
-        maxMsgLength = context.getInteger(ConfigConstants.MAX_MSG_LENGTH, MAX_MSG_DEFAULT_LENGTH);
-        Preconditions.checkArgument(
-                (maxMsgLength >= MSG_MIN_LENGTH && maxMsgLength <= ConfigConstants.MSG_MAX_LENGTH_BYTES),
-                "maxMsgLength must be >= 4 and <= " + ConfigConstants.MSG_MAX_LENGTH_BYTES);
-        isCompressed = context.getBoolean(ConfigConstants.MSG_COMPRESSED, true);
-
-        filterEmptyMsg = context.getBoolean(ConfigConstants.FILTER_EMPTY_MSG, false);
-
-        topic = context.getString(ConfigConstants.TOPIC, "");
-        attr = context.getString(ConfigConstants.ATTR);
-        Configurables.ensureRequiredNonNull(context, ConfigConstants.ATTR);
-
-        topic = topic.trim();
-        attr = attr.trim();
-        Preconditions.checkArgument(!attr.isEmpty(), "attr is empty");
-
-        statIntervalSec = context.getInteger(ConfigConstants.STAT_INTERVAL_SEC, INTERVAL_SEC);
-        Preconditions.checkArgument((statIntervalSec >= STAT_INTERVAL_MUST_THAN), "statIntervalSec must be >= 0");
-
-        pkgTimeoutSec = context.getInteger(ConfigConstants.PACKAGE_TIMEOUT_SEC, PKG_TIMEOUT_DEFAULT_SEC);
-
-        try {
-            maxConnections = context.getInteger(CONNECTIONS, DEFAULT_MAX_CONNECTIONS);
-        } catch (NumberFormatException e) {
-            logger.warn("BaseSource\'s \"connections\" property must specify an integer value.",
-                    context.getString(CONNECTIONS));
+    public void update() {
+        // check current all links
+        if (ConfigManager.getInstance().needChkIllegalIP()) {
+            int cnt = 0;
+            Channel channel;
+            String strRemoteIP;
+            long startTime = System.currentTimeMillis();
+            Iterator<Channel> iterator = allChannels.iterator();
+            while (iterator.hasNext()) {
+                channel = iterator.next();
+                strRemoteIP = AddressUtils.getChannelRemoteIP(channel);
+                if (strRemoteIP == null) {
+                    continue;
+                }
+                if (ConfigManager.getInstance().isIllegalIP(strRemoteIP)) {
+                    channel.disconnect();
+                    channel.close();
+                    allChannels.remove(channel);
+                    cnt++;
+                    logger.error(strRemoteIP + " is Illegal IP, so disconnect it !");
+                }
+            }
+            logger.info("Source {} channel check, disconnects {} Illegal channels, waist {} ms",
+                    getName(), cnt, (System.currentTimeMillis() - startTime));
         }
-
-        try {
-            maxThreads = context.getInteger(ConfigConstants.MAX_THREADS, DEFAULT_MAX_THREADS);
-        } catch (NumberFormatException e) {
-            logger.warn("Simple TCP Source max-threads property must specify an integer value. {}",
-                    context.getString(ConfigConstants.MAX_THREADS));
-        }
-        // get maxMonitorCnt's configure value
-        try {
-            maxMonitorCnt = context.getInteger(
-                    ConfigConstants.MAX_MONITOR_CNT, ConfigConstants.DEF_MONITOR_STAT_CNT);
-        } catch (NumberFormatException e) {
-            logger.warn("Property {} must specify an integer value: {}",
-                    ConfigConstants.MAX_MONITOR_CNT, context.getString(ConfigConstants.MAX_MONITOR_CNT));
-        }
-        Preconditions.checkArgument(maxMonitorCnt >= 0, "maxMonitorCnt must be >= 0");
-
-        receiveBufferSize = context.getInteger(ConfigConstants.RECEIVE_BUFFER_SIZE, RECEIVE_BUFFER_DEFAULT_SIZE);
-        if (receiveBufferSize > RECEIVE_BUFFER_MAX_SIZE) {
-            receiveBufferSize = RECEIVE_BUFFER_MAX_SIZE;
-        }
-        Preconditions.checkArgument(receiveBufferSize > BUFFER_SIZE_MUST_THAN,
-                "receiveBufferSize must be > 0");
-
-        sendBufferSize = context.getInteger(ConfigConstants.SEND_BUFFER_SIZE, SEND_BUFFER_DEFAULT_SIZE);
-        if (sendBufferSize > SEND_BUFFER_MAX_SIZE) {
-            sendBufferSize = SEND_BUFFER_MAX_SIZE;
-        }
-        Preconditions.checkArgument(sendBufferSize > BUFFER_SIZE_MUST_THAN,
-                "sendBufferSize must be > 0");
-
-        enableBusyWait = context.getBoolean(ConfigConstants.ENABLE_BUSY_WAIT, false);
-
-        this.customProcessor = context.getBoolean(ConfigConstants.CUSTOM_CHANNEL_PROCESSOR, false);
-    }
-
-    /**
-     * channel factory
-     * @return
-     */
-    public ChannelInitializer getChannelInitializerFactory() {
-        logger.info(new StringBuffer("load msgFactory=").append(msgFactoryName)
-                .append(" and serviceDecoderName=").append(serviceDecoderName).toString());
-        ChannelInitializer fac = null;
-        try {
-            ServiceDecoder serviceDecoder = (ServiceDecoder) Class.forName(serviceDecoderName).newInstance();
-            Class<? extends ChannelInitializer> clazz =
-                    (Class<? extends ChannelInitializer>) Class.forName(msgFactoryName);
-            Constructor ctor = clazz.getConstructor(BaseSource.class, ChannelGroup.class,
-                    String.class, ServiceDecoder.class, String.class, Integer.class,
-                    String.class, String.class, Boolean.class,
-                    Integer.class, Boolean.class, MonitorIndex.class,
-                    MonitorIndexExt.class, String.class);
-            logger.info("Using channel processor:{}", getChannelProcessor().getClass().getName());
-            fac = (ChannelInitializer) ctor.newInstance(this, allChannels,
-                    this.getProtocolName(), serviceDecoder, messageHandlerName, maxMsgLength,
-                    topic, attr, filterEmptyMsg,
-                    maxConnections, isCompressed, monitorIndex,
-                    monitorIndexExt, this.getProtocolName());
-        } catch (Exception e) {
-            logger.error(
-                    "Simple {} Source start error, fail to construct ChannelPipelineFactory with name "
-                            + "{}, ex {}",
-                    this.getProtocolName(), msgFactoryName, e);
-            stop();
-            throw new FlumeException(e.getMessage());
-        }
-        return fac;
     }
 
     /**
      * get metricItemSet
+     *
      * @return the metricItemSet
      */
     public DataProxyMetricItemSet getMetricItemSet() {
@@ -358,8 +349,232 @@ public abstract class BaseSource
         return context;
     }
 
+    public String getSrcHost() {
+        return srcHost;
+    }
+
+    public int getSrcPort() {
+        return srcPort;
+    }
+
+    public String getStrPort() {
+        return strPort;
+    }
+
+    public String getDefAttr() {
+        return defAttr;
+    }
+
+    public int getMaxMsgLength() {
+        return maxMsgLength;
+    }
+
+    public boolean isCompressed() {
+        return isCompressed;
+    }
+
+    public boolean isFilterEmptyMsg() {
+        return filterEmptyMsg;
+    }
+
+    public boolean isCustomProcessor() {
+        return customProcessor;
+    }
+
+    public int getMaxConnections() {
+        return maxConnections;
+    }
+
+    public ChannelGroup getAllChannels() {
+        return allChannels;
+    }
+
+    public long getMaxReadIdleTimeMs() {
+        return maxReadIdleTimeMs;
+    }
+
+    public String getMessageHandlerName() {
+        return messageHandlerName;
+    }
+
+    public int getMaxWorkerThreads() {
+        return maxWorkerThreads;
+    }
+
+    public void fileMetricIncSumStats(String eventKey) {
+        if (CommonConfigHolder.getInstance().isEnableFileMetric()) {
+            monitorStats.incSumStats(eventKey);
+        }
+    }
+
+    public void fileMetricIncDetailStats(String eventKey) {
+        if (CommonConfigHolder.getInstance().isEnableFileMetric()) {
+            monitorStats.incDetailStats(eventKey);
+        }
+    }
+
+    public void fileMetricAddSuccCnt(String key, int cnt, int packCnt, long packSize) {
+        if (CommonConfigHolder.getInstance().isEnableFileMetric()) {
+            monitorIndex.addSuccStats(key, cnt, packCnt, packSize);
+        }
+    }
+
+    public void fileMetricAddFailCnt(String key, int failCnt) {
+        if (CommonConfigHolder.getInstance().isEnableFileMetric()) {
+            monitorIndex.addFailStats(key, failCnt);
+        }
+    }
+    /**
+     * addMetric
+     *
+     * @param result
+     * @param size
+     * @param event
+     */
+    public void addMetric(boolean result, long size, Event event) {
+        Map<String, String> dimensions = new HashMap<>();
+        dimensions.put(DataProxyMetricItem.KEY_CLUSTER_ID, CommonConfigHolder.getInstance().getClusterName());
+        dimensions.put(DataProxyMetricItem.KEY_SOURCE_ID, getName());
+        dimensions.put(DataProxyMetricItem.KEY_SOURCE_DATA_ID, getStrPort());
+        DataProxyMetricItem.fillInlongId(event, dimensions);
+        DataProxyMetricItem.fillAuditFormatTime(event, dimensions);
+        DataProxyMetricItem metricItem = metricItemSet.findMetricItem(dimensions);
+        if (result) {
+            metricItem.readSuccessCount.incrementAndGet();
+            metricItem.readSuccessSize.addAndGet(size);
+            AuditUtils.add(AuditUtils.AUDIT_ID_DATAPROXY_READ_SUCCESS, event);
+        } else {
+            metricItem.readFailCount.incrementAndGet();
+            metricItem.readFailSize.addAndGet(size);
+        }
+    }
+
+    /**
+     * channel factory
+     *
+     * @return
+     */
+    public ChannelInitializer getChannelInitializerFactory() {
+        ChannelInitializer fac = null;
+        logger.info(this.getName() + " load msgFactory=" + msgFactoryName);
+        try {
+            Class<? extends ChannelInitializer> clazz =
+                    (Class<? extends ChannelInitializer>) Class.forName(msgFactoryName);
+            Constructor ctor = clazz.getConstructor(BaseSource.class);
+            logger.info("Using channel processor:{}", getChannelProcessor().getClass().getName());
+            fac = (ChannelInitializer) ctor.newInstance(this);
+        } catch (Exception e) {
+            logger.error("{} start error, fail to construct ChannelPipelineFactory with name {}",
+                    this.getName(), msgFactoryName, e);
+            stop();
+            throw new FlumeException(e.getMessage());
+        }
+        return fac;
+    }
+
     public abstract String getProtocolName();
 
     public abstract void startSource();
+
+    /**
+     * stopService
+     */
+    @Override
+    public void stopService() {
+        this.isRejectService = true;
+    }
+
+    /**
+     * recoverService
+     */
+    @Override
+    public void recoverService() {
+        this.isRejectService = false;
+    }
+
+    /**
+     * isRejectService
+     *
+     * @return
+     */
+    public boolean isRejectService() {
+        return isRejectService;
+    }
+
+    /**
+     * getHostIp
+     *
+     * @param context
+     * @return
+     */
+    private String getHostIp(Context context) {
+        String result = null;
+        // first get host ip from dataProxy.conf
+        String tmpVal = context.getString(SourceConstants.SRCCXT_CONFIG_HOST);
+        if (StringUtils.isNotBlank(tmpVal)) {
+            tmpVal = tmpVal.trim();
+            Preconditions.checkArgument(ConfStringUtils.isValidIp(tmpVal),
+                    SourceConstants.SRCCXT_CONFIG_HOST + "(" + tmpVal + ") config in conf not valid");
+            result = tmpVal;
+        }
+        // second get host ip from system env
+        Map<String, String> envMap = System.getenv();
+        if (envMap.containsKey(SourceConstants.SYSENV_HOST_IP)) {
+            tmpVal = envMap.get(SourceConstants.SYSENV_HOST_IP);
+            Preconditions.checkArgument(ConfStringUtils.isValidIp(tmpVal),
+                    SourceConstants.SYSENV_HOST_IP + "(" + tmpVal + ") config in system env not valid");
+            result = tmpVal.trim();
+        }
+        if (StringUtils.isBlank(result)) {
+            result = SourceConstants.VAL_DEF_HOST_VALUE;
+        }
+        return result;
+    }
+
+    /**
+     * getHostPort
+     *
+     * @param context
+     * @return
+     */
+    private int getHostPort(Context context) {
+        Integer result = null;
+        // first get host port from dataProxy.conf
+        String tmpVal = context.getString(SourceConstants.SRCCXT_CONFIG_PORT);
+        if (StringUtils.isNotBlank(tmpVal)) {
+            tmpVal = tmpVal.trim();
+            try {
+                result = Integer.parseInt(tmpVal);
+            } catch (Throwable e) {
+                throw new IllegalArgumentException(
+                        SourceConstants.SYSENV_HOST_PORT + "(" + tmpVal + ") config in conf not integer");
+            }
+        }
+        if (result != null) {
+            Preconditions.checkArgument(ConfStringUtils.isValidPort(result),
+                    SourceConstants.SRCCXT_CONFIG_PORT + "(" + result + ") config in conf not valid");
+        }
+        // second get host port from system env
+        Map<String, String> envMap = System.getenv();
+        if (envMap.containsKey(SourceConstants.SYSENV_HOST_PORT)) {
+            tmpVal = envMap.get(SourceConstants.SYSENV_HOST_PORT);
+            if (StringUtils.isNotBlank(tmpVal)) {
+                tmpVal = tmpVal.trim();
+                try {
+                    result = Integer.parseInt(tmpVal);
+                } catch (Throwable e) {
+                    throw new IllegalArgumentException(
+                            SourceConstants.SYSENV_HOST_PORT + "(" + tmpVal + ") config in system env not integer");
+                }
+                Preconditions.checkArgument(ConfStringUtils.isValidPort(result),
+                        SourceConstants.SYSENV_HOST_PORT + "(" + tmpVal + ") config in system env not valid");
+            }
+        }
+        if (result == null) {
+            throw new IllegalArgumentException("Required parameter " +
+                    SourceConstants.SRCCXT_CONFIG_PORT + " must exist and may not be null");
+        }
+        return result;
+    }
 
 }
