--- inlong-dataproxy/dataproxy-source/src/main/java/org/apache/inlong/dataproxy/source/ServiceDecoder.java
@@ -1,41 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.inlong.dataproxy.source;
-
-import io.netty.buffer.ByteBuf;
-import io.netty.channel.Channel;
-
-import java.util.Map;
-
-public interface ServiceDecoder {
-
-    int HEAD_LENGTH = 4;
-
-    /**
-     * extract data from buffer and convert it into map.
-     * 
-     * @param cb           the message Byte buffer
-     * @param strRemoteIP  the remote ip message sent
-     * @param msgRcvTime   the received message time
-     * @param channel      the channel
-     * @return Map         the message map
-     * @throws Exception
-     */
-    Map<String, Object> extractData(ByteBuf cb, String strRemoteIP,
-            long msgRcvTime, Channel channel) throws Exception;
-}
