+++ inlong-dataproxy/dataproxy-source/src/main/java/org/apache/inlong/dataproxy/source/SimpleTcpSource.java
@@ -18,56 +18,33 @@
 package org.apache.inlong.dataproxy.source;
 
 import org.apache.inlong.dataproxy.config.ConfigManager;
-import org.apache.inlong.dataproxy.config.holder.ConfigUpdateCallback;
-import org.apache.inlong.dataproxy.consts.ConfigConstants;
-import org.apache.inlong.dataproxy.utils.AddressUtils;
+import org.apache.inlong.dataproxy.utils.ConfStringUtils;
 import org.apache.inlong.dataproxy.utils.EventLoopUtil;
 
 import com.google.common.base.Preconditions;
 import io.netty.bootstrap.ServerBootstrap;
 import io.netty.buffer.ByteBufAllocator;
-import io.netty.channel.Channel;
-import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelOption;
 import io.netty.util.concurrent.DefaultThreadFactory;
 import org.apache.flume.Context;
-import org.apache.flume.EventDrivenSource;
 import org.apache.flume.conf.Configurable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.InetSocketAddress;
-import java.util.Iterator;
 
 /**
  * Simple tcp source
- *
  */
-public class SimpleTcpSource extends BaseSource
-        implements
-            Configurable,
-            ConfigUpdateCallback,
-            EventDrivenSource {
+public class SimpleTcpSource extends BaseSource implements Configurable {
 
     private static final Logger logger = LoggerFactory.getLogger(SimpleTcpSource.class);
 
-    private static int TRAFFIC_CLASS_TYPE_0 = 0;
-
-    private static int TRAFFIC_CLASS_TYPE_96 = 96;
-
-    private static int HIGH_WATER_MARK_DEFAULT_VALUE = 64 * 1024;
-
-    private boolean tcpNoDelay = true;
-
-    private boolean keepAlive = true;
-
-    private int highWaterMark;
-
-    private int trafficClass;
-
-    protected String topic;
-
     private ServerBootstrap bootstrap;
+    private boolean tcpNoDelay;
+    private boolean tcpKeepAlive;
+    private int highWaterMark;
+    private boolean enableBusyWait;
 
     public SimpleTcpSource() {
         super();
@@ -75,51 +52,66 @@ public class SimpleTcpSource extends BaseSource
     }
 
     @Override
+    public void configure(Context context) {
+        logger.info("Source {} context is {}", getName(), context);
+        super.configure(context);
+        // get tcp no-delay parameter
+        this.tcpNoDelay = context.getBoolean(SourceConstants.SRCCXT_TCP_NO_DELAY,
+                SourceConstants.VAL_DEF_TCP_NO_DELAY);
+        // get tcp keep-alive parameter
+        this.tcpKeepAlive = context.getBoolean(SourceConstants.SRCCXT_TCP_KEEP_ALIVE,
+                SourceConstants.VAL_DEF_TCP_KEEP_ALIVE);
+        // get tcp enable busy-wait
+        this.enableBusyWait = context.getBoolean(SourceConstants.SRCCXT_TCP_ENABLE_BUSY_WAIT,
+                SourceConstants.VAL_DEF_TCP_ENABLE_BUSY_WAIT);
+        // get tcp high watermark
+        this.highWaterMark = ConfStringUtils.getIntValue(context,
+                SourceConstants.SRCCXT_TCP_HIGH_WATER_MARK, SourceConstants.VAL_DEF_TCP_HIGH_WATER_MARK);
+        Preconditions.checkArgument((this.highWaterMark >= SourceConstants.VAL_MIN_TCP_HIGH_WATER_MARK),
+                SourceConstants.VAL_DEF_TCP_HIGH_WATER_MARK + " must be >= "
+                        + SourceConstants.VAL_MIN_TCP_HIGH_WATER_MARK);
+    }
+
+    @Override
     public synchronized void startSource() {
         logger.info("start " + this.getName());
-
-        logger.info("Set max workers : {} ;", maxThreads);
-
-        acceptorThreadFactory = new DefaultThreadFactory("tcpSource-nettyBoss-threadGroup");
-
-        this.acceptorGroup = EventLoopUtil.newEventLoopGroup(
-                acceptorThreads, false, acceptorThreadFactory);
-
-        this.workerGroup = EventLoopUtil
-                .newEventLoopGroup(maxThreads, enableBusyWait,
-                        new DefaultThreadFactory("tcpSource-nettyWorker-threadGroup"));
-
+        // build accept group
+        this.acceptorGroup = EventLoopUtil.newEventLoopGroup(maxAcceptThreads, enableBusyWait,
+                new DefaultThreadFactory(this.getName() + "-boss-group"));
+        // build worker group
+        this.workerGroup = EventLoopUtil.newEventLoopGroup(maxWorkerThreads, enableBusyWait,
+                new DefaultThreadFactory(this.getName() + "-worker-group"));
+        // init boostrap
         bootstrap = new ServerBootstrap();
-
+        if (conLinger >= 0) {
+            bootstrap.option(ChannelOption.SO_LINGER, conLinger);
+        }
+        bootstrap.option(ChannelOption.SO_BACKLOG, conBacklog);
+        bootstrap.option(ChannelOption.SO_REUSEADDR, reuseAddress);
         bootstrap.childOption(ChannelOption.ALLOCATOR, ByteBufAllocator.DEFAULT);
         bootstrap.childOption(ChannelOption.TCP_NODELAY, tcpNoDelay);
-        bootstrap.childOption(ChannelOption.SO_KEEPALIVE, keepAlive);
-        bootstrap.childOption(ChannelOption.SO_RCVBUF, receiveBufferSize);
-        bootstrap.childOption(ChannelOption.SO_SNDBUF, sendBufferSize);
-        // serverBootstrap.childOption("child.trafficClass", trafficClass);
+        bootstrap.childOption(ChannelOption.SO_KEEPALIVE, tcpKeepAlive);
+        bootstrap.childOption(ChannelOption.SO_RCVBUF, maxRcvBufferSize);
+        bootstrap.childOption(ChannelOption.SO_SNDBUF, maxSendBufferSize);
         bootstrap.childOption(ChannelOption.WRITE_BUFFER_HIGH_WATER_MARK, highWaterMark);
         bootstrap.channel(EventLoopUtil.getServerSocketChannelClass(workerGroup));
         EventLoopUtil.enableTriggeredMode(bootstrap);
         bootstrap.group(acceptorGroup, workerGroup);
-        logger.info("load msgFactory=" + msgFactoryName
-                + " and serviceDecoderName=" + serviceDecoderName);
-
-        ChannelInitializer fac = this.getChannelInitializerFactory();
-        bootstrap.childHandler(fac);
+        bootstrap.childHandler(this.getChannelInitializerFactory());
         try {
-            if (host == null) {
-                channelFuture = bootstrap.bind(new InetSocketAddress(port)).sync();
+            if (srcHost == null) {
+                channelFuture = bootstrap.bind(new InetSocketAddress(srcPort)).sync();
             } else {
-                channelFuture = bootstrap.bind(new InetSocketAddress(host, port)).sync();
+                channelFuture = bootstrap.bind(new InetSocketAddress(srcHost, srcPort)).sync();
             }
         } catch (Exception e) {
-            logger.error("Simple TCP Source error bind host {} port {},program will exit! e = {}",
-                    host, port, e);
+            logger.error("Source {} bind ({}:{}) error, program will exit! e = {}",
+                    this.getName(), srcHost, srcPort, e);
             System.exit(-1);
         }
         ConfigManager.getInstance().addSourceReportInfo(
-                host, String.valueOf(port), getProtocolName().toUpperCase());
-        logger.info("Simple TCP Source started at host {}, port {}", host, port);
+                srcHost, String.valueOf(srcPort), getProtocolName().toUpperCase());
+        logger.info("Source {} started at ({}:{})!", this.getName(), srcHost, srcPort);
     }
 
     @Override
@@ -128,70 +120,8 @@ public class SimpleTcpSource extends BaseSource
     }
 
     @Override
-    public void configure(Context context) {
-        logger.info("context is {}", context);
-        super.configure(context);
-        tcpNoDelay = context.getBoolean(ConfigConstants.TCP_NO_DELAY, true);
-        keepAlive = context.getBoolean(ConfigConstants.KEEP_ALIVE, true);
-        highWaterMark = context.getInteger(ConfigConstants.HIGH_WATER_MARK, HIGH_WATER_MARK_DEFAULT_VALUE);
-        receiveBufferSize = context.getInteger(ConfigConstants.RECEIVE_BUFFER_SIZE, RECEIVE_BUFFER_DEFAULT_SIZE);
-        if (receiveBufferSize > RECEIVE_BUFFER_MAX_SIZE) {
-            receiveBufferSize = RECEIVE_BUFFER_MAX_SIZE;
-        }
-        Preconditions.checkArgument(receiveBufferSize > BUFFER_SIZE_MUST_THAN,
-                "receiveBufferSize must be > 0");
-
-        sendBufferSize = context.getInteger(ConfigConstants.SEND_BUFFER_SIZE, SEND_BUFFER_DEFAULT_SIZE);
-        if (sendBufferSize > SEND_BUFFER_MAX_SIZE) {
-            sendBufferSize = SEND_BUFFER_MAX_SIZE;
-        }
-        Preconditions.checkArgument(sendBufferSize > BUFFER_SIZE_MUST_THAN,
-                "sendBufferSize must be > 0");
-
-        trafficClass = context.getInteger(ConfigConstants.TRAFFIC_CLASS, TRAFFIC_CLASS_TYPE_0);
-        Preconditions.checkArgument((trafficClass == TRAFFIC_CLASS_TYPE_0
-                || trafficClass == TRAFFIC_CLASS_TYPE_96),
-                "trafficClass must be == 0 or == 96");
-
-        try {
-            maxThreads = context.getInteger(ConfigConstants.MAX_THREADS, 32);
-        } catch (NumberFormatException e) {
-            logger.warn("Simple TCP Source max-threads property must specify an integer value. {}",
-                    context.getString(ConfigConstants.MAX_THREADS));
-        }
-    }
-
-    @Override
     public String getProtocolName() {
-        return "tcp";
-    }
-
-    @Override
-    public void update() {
-        // check current all links
-        if (ConfigManager.getInstance().needChkIllegalIP()) {
-            int cnt = 0;
-            Channel channel;
-            String strRemoteIP;
-            long startTime = System.currentTimeMillis();
-            Iterator<Channel> iterator = allChannels.iterator();
-            while (iterator.hasNext()) {
-                channel = iterator.next();
-                strRemoteIP = AddressUtils.getChannelRemoteIP(channel);
-                if (strRemoteIP == null) {
-                    continue;
-                }
-                if (ConfigManager.getInstance().isIllegalIP(strRemoteIP)) {
-                    channel.disconnect();
-                    channel.close();
-                    allChannels.remove(channel);
-                    cnt++;
-                    logger.error(strRemoteIP + " is Illegal IP, so disconnect it !");
-                }
-            }
-            logger.info("Channel check, {} disconnects {} Illegal channels, waist {} ms",
-                    getName(), cnt, (System.currentTimeMillis() - startTime));
-        }
+        return SourceConstants.SRC_PROTOCOL_TYPE_TCP;
     }
 
 }
